Update code to address lint / analysis warnings. Remove unused resources and classes.

+++ app/src/main/java/co/nano/nanowallet/ui/intro/IntroSeedFragment.java
@@ -90,7 +90,7 @@ public class IntroSeedFragment extends BaseFragment {
             int keypadHeight = screenHeight - r.bottom;
 
             if (keypadHeight > screenHeight * 0.15) { // 0.15 ratio is perhaps enough to determine keypad height.
-                binding.introSeedScrollview.smoothScrollTo(0, (int) (binding.introSeedScrollview.getMaxScrollAmount()));
+                binding.introSeedScrollview.smoothScrollTo(0, binding.introSeedScrollview.getMaxScrollAmount());
             }
         });
